+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 - EntityTableModel.deleteSelected() now returns the deleted entities.
 ### is.codion.swing.framework.ui
 - EntityPanel.DetailLayout.layout(), entityPanel parameter removed, now returns Optional, related changes.
+- EntityPanel.DetailController reintroduced, related changes.
 
 ## 0.17.35
 ### is.codion.common.model